Rename TimelineDateGroup to MonthSegment
